Gold tier/requirements.txt
# AI Employee — Gold Tier Requirements
# Install with: pip install -r requirements.txt
# After install, also run: playwright install chromium

# ── File system monitoring ───────────────────────────────────
watchdog>=4.0.0

# ── Gmail Watcher ────────────────────────────────────────────
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# ── WhatsApp & LinkedIn (Playwright) ────────────────────────
playwright>=1.40.0

# ── Facebook & Instagram (Graph API) ────────────────────────
requests>=2.31.0

# ── Twitter/X (API v2) ──────────────────────────────────────
requests-oauthlib>=1.3.0

# ── Odoo (JSON-RPC client) ──────────────────────────────────
# Uses requests (already listed above)

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
